+++ tests/pytest/test_projects.py
@@ -32,12 +32,18 @@ class TestProjects(TestCase):
         # idempotent - do it again
         result = self.run_cli(cmd_env, sub_cmd + f"set {proj_name} --desc \"{new_desc}\"")
         self.assertEqual(result.return_value, 0)
-        self.assertTrue(result.err_contains_value(f"Project '{proj_name}' not updated: same description"))
+
+        # rename
+        orig_name = proj_name
+        proj_name = self.make_name("test-proj-rename")
+        result = self.run_cli(cmd_env, sub_cmd + f"set {orig_name} --rename \"{proj_name}\"")
+        self.assertEqual(result.return_value, 0)
+        self.assertIn(f"Updated project '{proj_name}'", result.out())
 
         # nothing to update
         result = self.run_cli(cmd_env, sub_cmd + f"set {proj_name}")
         self.assertEqual(result.return_value, 0)
-        self.assertTrue(result.err_contains_value(f"Project '{proj_name}' not updated: no description"))
+        self.assertTrue(result.err_contains_value(f"Project '{proj_name}' not updated: no updated parameters provided"))
 
         # test the list without the table
         result = self.run_cli(cmd_env, sub_cmd + "list")
@@ -51,7 +57,7 @@ class TestProjects(TestCase):
         self.assertTrue(result.out_contains_value(proj_name))
         self.assertTrue(result.out_contains_both(proj_name, new_desc))
 
-        # delete the description
+        # delete
         result = self.run_cli(cmd_env, sub_cmd + f"delete {proj_name} --confirm")
         self.assertEqual(result.return_value, 0)
         result = self.run_cli(cmd_env, sub_cmd + "ls -v")